+++ stories/button.tsx
@@ -36,14 +36,11 @@ const Comp: React.FC<CompProps> = props => {
 }
 
 const requestData: RequestData = {
+  version: 1,
   action: Action.attestation,
   token: 'a08714b92346a1bba4262ed575d23de3ff3e6b5480ad0e1c82c011bab0411fdf',
   url: 'https://receive-kit.bloom.co/api/receive',
-  org_logo_url: 'https://bloom.co/images/notif/bloom-logo.png',
-  org_name: 'Bloom',
-  org_usage_policy_url: 'https://bloom.co/legal/terms',
-  org_privacy_policy_url: 'https://bloom.co/legal/privacy',
-  types: ['phone', 'email'],
+  payload_url: 'https://receive-kit.bloom.co/api/get-payload?token=',
 }
 
 const Updating: React.FC = () => {