+++ test/elements/renderRequestButton.test.ts
@@ -27,14 +27,11 @@ describe('renderRequestButton', () => {
     requestButton = renderRequestButton({
       container,
       requestData: {
+        version: 1,
         action: Action.attestation,
         token: 'token',
         url: 'https://receive-kit.bloom.co/api/receive',
-        org_logo_url: 'https://bloom.co/images/notif/bloom-logo.png',
-        org_name: 'Bloom',
-        org_usage_policy_url: 'https://bloom.co/legal/terms',
-        org_privacy_policy_url: 'https://bloom.co/legal/privacy',
-        types: ['phone', 'email'],
+        payload_url: 'https://receive-kit.bloom.co/api/get-payload',
       },
       buttonOptions: {
         callbackUrl: 'https://bloom.co/callback-url',
@@ -62,14 +59,11 @@ describe('renderRequestButton', () => {
   test('renders a medium button', () => {
     requestButton.update({
       requestData: {
+        version: 1,
         action: Action.attestation,
         token: 'token',
         url: 'https://receive-kit.bloom.co/api/receive',
-        org_logo_url: 'https://bloom.co/images/notif/bloom-logo.png',
-        org_name: 'Bloom',
-        org_usage_policy_url: 'https://bloom.co/legal/terms',
-        org_privacy_policy_url: 'https://bloom.co/legal/privacy',
-        types: ['phone', 'email'],
+        payload_url: 'https://receive-kit.bloom.co/api/get-payload',
       },
       buttonOptions: {
         callbackUrl: 'https://bloom.co/callback-url',
@@ -83,14 +77,11 @@ describe('renderRequestButton', () => {
   test('renders a small button', () => {
     requestButton.update({
       requestData: {
+        version: 1,
         action: Action.attestation,
         token: 'token',
         url: 'https://receive-kit.bloom.co/api/receive',
-        org_logo_url: 'https://bloom.co/images/notif/bloom-logo.png',
-        org_name: 'Bloom',
-        org_usage_policy_url: 'https://bloom.co/legal/terms',
-        org_privacy_policy_url: 'https://bloom.co/legal/privacy',
-        types: ['phone', 'email'],
+        payload_url: 'https://receive-kit.bloom.co/api/get-payload',
       },
       buttonOptions: {
         callbackUrl: 'https://bloom.co/callback-url',
@@ -105,14 +96,11 @@ describe('renderRequestButton', () => {
   test('updates the button', () => {
     requestButton.update({
       requestData: {
+        version: 1,
         action: Action.attestation,
-        token: 'token 2',
+        token: 'token',
         url: 'https://receive-kit.bloom.co/api/receive',
-        org_logo_url: 'https://bloom.co/images/notif/bloom-logo.png',
-        org_name: 'Bloom',
-        org_usage_policy_url: 'https://bloom.co/legal/terms',
-        org_privacy_policy_url: 'https://bloom.co/legal/privacy',
-        types: ['phone', 'email'],
+        payload_url: 'https://receive-kit.bloom.co/api/get-payload',
       },
       buttonOptions: {
         callbackUrl: 'https://bloom.co/callback-url-2',